+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/models/__init__.py
@@ -37,6 +37,7 @@ try:
     from ._models_py3 import MetricSpecification
     from ._models_py3 import NameAvailability
     from ._models_py3 import NameAvailabilityParameters
+    from ._models_py3 import NetworkProfile
     from ._models_py3 import OperationDetail
     from ._models_py3 import OperationDisplay
     from ._models_py3 import OperationProperties
@@ -88,6 +89,7 @@ except (SyntaxError, ImportError):
     from ._models import MetricSpecification
     from ._models import NameAvailability
     from ._models import NameAvailabilityParameters
+    from ._models import NetworkProfile
     from ._models import OperationDetail
     from ._models import OperationDisplay
     from ._models import OperationProperties
@@ -163,6 +165,7 @@ __all__ = [
     'MetricSpecification',
     'NameAvailability',
     'NameAvailabilityParameters',
+    'NetworkProfile',
     'OperationDetail',
     'OperationDisplay',
     'OperationProperties',