+++ src/synapse/HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ===============
 
+0.3.0
+++++++
+* Fix a bug when running pyspark session statement.
+
+0.2.0
+++++++
+* Support resource id parameter when getting, updating or deleting synapse Workspace, Spark pool and SQL pool.
+* Add command group `az synapse workspace firewall-rule` to add and delete firewall rule.
+* Add a command `az syanpse workspace check-name` to check a workspace exist or not.
+* Refactor the parameters'name under command group `az synapse spark` and `az synapse sql`.
+
 0.1.0
 ++++++
 * Initial release.